Improve homepage CSS
Before, the homepage used a hand-crafted minimal CSS for the Bootstrap
header and search form. Now, the original Bootstrap CSS file remains
untouched and instead the homepage HTML is stripped from certain CSS
ids and classes using a script. This approach is more robust and avoids
having to load a custom Bootstrap CSS file that's only used on the
homepage.